requirements.txt
Flask==3.0.3
scikit-learn==1.4.2
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
